+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `hideTax` prop on OrderTotal
+
 ## [2.8.2] - 2021-03-03
 
 ### Fixed